fix: layout spacing, detekt violations, and class size for CI
- Fix all auth page layouts: centered viewport, proper DaisyUI tabs,
  full-width inputs, btn-primary submit buttons, checkbox-sm, divider
- Replace all inline styles with DaisyUI/Tailwind utility classes
- Remove stale CSS class names (button-link, message-meta, topbar-label,
  pagination-info) replaced by DaisyUI equivalents
- Use DaisyUI dialog/modal for delete confirmation in ProfilePage
- Fix buildUponDefaultConfig:true in pom.xml (was false)
- Fix missing assertNotNull import in SwingAppE2ETest
- Fix ImplicitDefaultLocale in SwingSyncApp Color.toHtml()
- Remove unused handle parameter from JdbiMessageRepository
- Split SyncWindow (1541→586 lines) into SyncDialogs, SyncProfilePanel,
  SyncViews to pass detekt LargeClass rule
- Fix LongParameterList in SyncViews with component grouping classes
