+++ platform-persistence-jdbi/src/main/kotlin/io/github/rygel/outerstellar/platform/persistence/JdbiMessageRepository.kt
@@ -21,7 +21,7 @@ class JdbiMessageRepository(private val jdbi: Jdbi) : MessageRepository {
         includeDeleted: Boolean,
     ): List<MessageSummary> {
         return jdbi.withHandle<List<MessageSummary>, Exception> { handle ->
-            val (whereClause, bindings) = buildFilterClause(handle, query, year, includeDeleted)
+            val (whereClause, bindings) = buildFilterClause(query, year, includeDeleted)
             val sql =
                 """
                 SELECT sync_id, author, content, updated_at_epoch_ms, dirty, deleted, version, sync_conflict
@@ -42,7 +42,7 @@ class JdbiMessageRepository(private val jdbi: Jdbi) : MessageRepository {
 
     override fun countMessages(query: String?, year: Int?, includeDeleted: Boolean): Long {
         return jdbi.withHandle<Long, Exception> { handle ->
-            val (whereClause, bindings) = buildFilterClause(handle, query, year, includeDeleted)
+            val (whereClause, bindings) = buildFilterClause(query, year, includeDeleted)
             val q = handle.createQuery("SELECT COUNT(*) FROM plt_messages WHERE $whereClause")
             bindings(q)
             q.mapTo(Long::class.java).one()
@@ -297,7 +297,7 @@ class JdbiMessageRepository(private val jdbi: Jdbi) : MessageRepository {
 
     private data class FilterClause(val sql: String, val binder: (org.jdbi.v3.core.statement.Query) -> Unit)
 
-    private fun buildFilterClause(handle: Handle, query: String?, year: Int?, includeDeleted: Boolean): FilterClause {
+    private fun buildFilterClause(query: String?, year: Int?, includeDeleted: Boolean): FilterClause {
         val conditions = mutableListOf<String>()
         val bindings = mutableMapOf<String, Any>()
