+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/ChangePasswordForm.kte
@@ -1,23 +1,29 @@
 @import io.github.rygel.outerstellar.platform.web.ChangePasswordForm
 @param model: ChangePasswordForm
 
-<form class="flex flex-col gap-3"
+<form class="flex flex-col gap-4"
       hx-post="${model.submitUrl}"
       hx-target="#auth-result-slot"
       hx-swap="innerHTML">
-    <h3>${model.title}</h3>
+    <h3 class="text-xl font-bold">${model.title}</h3>
 
-    <label for="currentPassword">${model.currentPasswordLabel}</label>
-    <input type="password" id="currentPassword" name="currentPassword"
-           placeholder="${model.currentPasswordPlaceholder}" required class="input input-bordered">
+    <div class="flex flex-col gap-1.5">
+        <label for="currentPassword" class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${model.currentPasswordLabel}</label>
+        <input type="password" id="currentPassword" name="currentPassword"
+               placeholder="${model.currentPasswordPlaceholder}" required class="input input-bordered w-full">
+    </div>
 
-    <label for="newPassword">${model.newPasswordLabel}</label>
-    <input type="password" id="newPassword" name="newPassword"
-           placeholder="${model.newPasswordPlaceholder}" required class="input input-bordered">
+    <div class="flex flex-col gap-1.5">
+        <label for="newPassword" class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${model.newPasswordLabel}</label>
+        <input type="password" id="newPassword" name="newPassword"
+               placeholder="${model.newPasswordPlaceholder}" required class="input input-bordered w-full">
+    </div>
 
-    <label for="confirmPassword">${model.confirmPasswordLabel}</label>
-    <input type="password" id="confirmPassword" name="confirmPassword"
-           placeholder="${model.confirmPasswordPlaceholder}" required class="input input-bordered">
+    <div class="flex flex-col gap-1.5">
+        <label for="confirmPassword" class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${model.confirmPasswordLabel}</label>
+        <input type="password" id="confirmPassword" name="confirmPassword"
+               placeholder="${model.confirmPasswordPlaceholder}" required class="input input-bordered w-full">
+    </div>
 
-    <button type="submit">${model.submitLabel}</button>
+    <button type="submit" class="btn btn-primary w-full mt-2">${model.submitLabel}</button>
 </form>
