+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/ChangePasswordPage.kte
@@ -3,12 +3,18 @@
 @param model: Page<ChangePasswordPage>
 
 @template.io.github.rygel.outerstellar.platform.web.LayoutRouter(shell = model.shell, content = @`
-    <div style="max-width: 500px; margin: 2rem auto;">
-        <section class="card bg-base-200 border border-base-300">
-            <div id="change-password-form">
-                @template.io.github.rygel.outerstellar.platform.web.ChangePasswordForm(model = model.data.form)
+    <div class="flex items-center justify-center min-h-[80vh]">
+        <div class="w-full max-w-md">
+            <div class="text-center mb-8">
+                <h1 class="text-3xl font-extrabold">${model.shell.pageTitle}</h1>
             </div>
-            <div id="auth-result-slot" class="mt-4"></div>
-        </section>
+
+            <section class="card bg-base-200 border border-base-300 p-8">
+                <div id="change-password-form">
+                    @template.io.github.rygel.outerstellar.platform.web.ChangePasswordForm(model = model.data.form)
+                </div>
+                <div id="auth-result-slot" class="mt-4"></div>
+            </section>
+        </div>
     </div>
 `)
