+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/ErrorPage.kte
@@ -9,20 +9,20 @@
             <h1>${model.data.heading}</h1>
             <p>${model.data.message}</p>
         </div>
-        <div class="flex gap-3 justify-end">
-            <a href="${model.data.primaryActionUrl}" class="button-link">${model.data.primaryActionLabel}</a>
-            <a href="${model.data.secondaryActionUrl}" class="button-link btn btn-ghost">${model.data.secondaryActionLabel}</a>
+        <div class="flex gap-3 justify-end mt-4">
+            <a href="${model.data.primaryActionUrl}" class="btn btn-primary btn-sm">${model.data.primaryActionLabel}</a>
+            <a href="${model.data.secondaryActionUrl}" class="btn btn-ghost btn-sm">${model.data.secondaryActionLabel}</a>
         </div>
     </section>
 
     <section class="card bg-base-200 border border-base-300">
-        <button class="button-link btn btn-ghost"
+        <button class="btn btn-ghost btn-sm"
                 hx-get="${model.data.helpUrl}"
                 hx-target="#error-help-slot"
                 hx-swap="innerHTML">
             ${model.data.helpButtonLabel}
         </button>
-        <div id="error-help-slot" class="card bg-base-100 mt-6" style="margin-top: 1.5rem; display: none;" :style="{ display: 'block' }">
+        <div id="error-help-slot" class="mt-4" style="display: none;">
             <!-- Content loaded via HTMX -->
         </div>
     </section>
