+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/HomePage.kte
@@ -31,7 +31,7 @@
             <input type="text" name="author" placeholder="${model.data.authorPlaceholder}" class="input input-bordered">
             <textarea name="content" placeholder="${model.data.contentPlaceholder}" required style="min-height: 120px;" class="textarea textarea-bordered"></textarea>
             <div class="flex justify-end">
-                <button type="submit">
+                <button type="submit" class="btn btn-primary">
                     <i id="submit-indicator" class="ri-send-plane-fill htmx-indicator"></i>
                     ${model.data.submitLabel}
                 </button>
