+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/NotificationsPage.kte
@@ -3,16 +3,16 @@
 @param model: Page<NotificationsPage>
 
 @template.io.github.rygel.outerstellar.platform.web.LayoutRouter(shell = model.shell, content = @`
-    <div style="max-width: 760px; margin: 0 auto;">
-        <div style="display:flex; align-items:center; justify-content:space-between; margin-bottom:1.5rem;">
+    <div class="max-w-3xl mx-auto">
+        <div class="flex items-center justify-between mb-6">
             <div>
-                <h2 style="margin:0;">${model.data.title}</h2>
-                <p style="margin:0.25rem 0 0; opacity:0.65; font-size:0.9rem;">${model.data.description}</p>
+                <h2 class="mt-0">${model.data.title}</h2>
+                <p class="opacity-60 text-sm mt-1">${model.data.description}</p>
             </div>
             @if(model.data.unreadCount > 0)
                 <form method="post" action="${model.data.markAllReadUrl}">
                     <input type="hidden" name="_csrf" value="${model.shell.csrfToken}">
-                    <button type="submit" class="btn btn-ghost" style="font-size:0.85rem;">
+                    <button type="submit" class="btn btn-ghost btn-sm">
                         <i class="ri-check-double-line"></i> ${model.data.markAllReadLabel}
                     </button>
                 </form>
@@ -20,46 +20,44 @@
         </div>
 
         @if(model.data.notifications.isEmpty())
-            <div class="card bg-base-100" style="text-align:center; padding:3rem;">
-                <i class="ri-notification-off-line" style="font-size:2.5rem; opacity:0.3;"></i>
-                <p style="margin:0.75rem 0 0; opacity:0.5;">${model.data.emptyLabel}</p>
+            <div class="card bg-base-100 text-center py-12">
+                <i class="ri-notification-off-line text-4xl opacity-30"></i>
+                <p class="opacity-50 mt-3">${model.data.emptyLabel}</p>
             </div>
         @else
-            <div style="display:flex; flex-direction:column; gap:0.75rem;">
+            <div class="flex flex-col gap-3">
                 @for(n in model.data.notifications)
-                    <div class="card@if(!n.read) bg-accent/10 border border-accent/30@else bg-base-200 border border-base-300@endif"
-                         style="display:flex; align-items:flex-start; gap:1rem; padding:1rem 1.25rem;">
-                        <div style="padding-top:0.15rem;">
+                    <div class="card@if(!n.read) bg-accent/10 border border-accent/30@else bg-base-200 border border-base-300@endif flex items-start gap-3 p-4">
+                        <div class="pt-0.5 shrink-0">
                             @if(n.type == "success")
-                                <i class="ri-checkbox-circle-line" style="color:var(--color-success, #22c55e); font-size:1.2rem;"></i>
+                                <i class="ri-checkbox-circle-line text-success text-xl"></i>
                             @elseif(n.type == "warning")
-                                <i class="ri-alert-line" style="color:var(--color-warning, #f59e0b); font-size:1.2rem;"></i>
+                                <i class="ri-alert-line text-warning text-xl"></i>
                             @elseif(n.type == "danger")
-                                <i class="ri-error-warning-line" style="color:var(--color-danger, #ef4444); font-size:1.2rem;"></i>
+                                <i class="ri-error-warning-line text-error text-xl"></i>
                             @else
-                                <i class="ri-information-line" style="color:var(--color-accent, #6366f1); font-size:1.2rem;"></i>
+                                <i class="ri-information-line text-accent text-xl"></i>
                             @endif
                         </div>
-                        <div style="flex:1; min-width:0;">
-                            <div style="display:flex; align-items:center; gap:0.5rem; flex-wrap:wrap;">
-                                <strong style="font-size:0.95rem;">${n.title}</strong>
+                        <div class="flex-1 min-w-0">
+                            <div class="flex items-center gap-2 flex-wrap">
+                                <strong class="text-sm">${n.title}</strong>
                                 @if(!n.read)
-                                    <span class="badge" style="font-size:0.7rem; padding:0.1rem 0.4rem;">${model.data.newLabel}</span>
+                                    <span class="badge badge-xs">${model.data.newLabel}</span>
                                 @endif
-                                <span style="opacity:0.45; font-size:0.8rem; margin-left:auto;">${n.timeAgo}</span>
+                                <span class="opacity-40 text-xs ml-auto">${n.timeAgo}</span>
                             </div>
-                            <p style="margin:0.25rem 0 0; font-size:0.875rem; opacity:0.8;">${n.body}</p>
+                            <p class="text-sm opacity-70 mt-0.5">${n.body}</p>
                         </div>
                         @if(!n.read)
-                            <form method="post" action="${n.markReadUrl}" style="flex-shrink:0;">
+                            <form method="post" action="${n.markReadUrl}" class="shrink-0">
                                 <input type="hidden" name="_csrf" value="${model.shell.csrfToken}">
-                                <button type="submit" class="btn btn-ghost" style="font-size:0.8rem; padding:0.3rem 0.6rem;"
-                                        title="${model.data.markReadLabel}">
+                                <button type="submit" class="btn btn-ghost btn-xs" title="${model.data.markReadLabel}">
                                     <i class="ri-check-line"></i>
                                 </button>
                             </form>
                         @else
-                            <span style="opacity:0.3; font-size:0.8rem; flex-shrink:0;">${model.data.readLabel}</span>
+                            <span class="opacity-30 text-xs shrink-0">${model.data.readLabel}</span>
                         @endif
                     </div>
                 @endfor
