+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/ResetPasswordPage.kte
@@ -3,29 +3,35 @@
 @param model: Page<ResetPasswordPage>
 
 @template.io.github.rygel.outerstellar.platform.web.LayoutRouter(shell = model.shell, content = @`
-    <div style="max-width: 500px; margin: 2rem auto;">
-        <section class="card bg-base-200 border border-base-300">
-            <div id="reset-password-form">
-                <form class="flex flex-col gap-3"
+    <div class="flex items-center justify-center min-h-[80vh]">
+        <div class="w-full max-w-md">
+            <div class="text-center mb-8">
+                <h1 class="text-3xl font-extrabold">${model.shell.pageTitle}</h1>
+            </div>
+
+            <section class="card bg-base-200 border border-base-300 p-8">
+                <form class="flex flex-col gap-4"
                       hx-post="${model.data.submitUrl}"
                       hx-target="#auth-result-slot"
                       hx-swap="innerHTML">
-                    <h3>${model.shell.pageTitle}</h3>
-
                     <input type="hidden" name="token" value="${model.data.token}">
 
-                    <label for="newPassword">${model.data.newPasswordLabel}</label>
-                    <input type="password" id="newPassword" name="newPassword"
-                           placeholder="${model.data.newPasswordPlaceholder}" required class="input input-bordered">
+                    <div class="flex flex-col gap-1.5">
+                        <label for="newPassword" class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${model.data.newPasswordLabel}</label>
+                        <input type="password" id="newPassword" name="newPassword"
+                               placeholder="${model.data.newPasswordPlaceholder}" required class="input input-bordered w-full">
+                    </div>
 
-                    <label for="confirmPassword">${model.data.confirmPasswordLabel}</label>
-                    <input type="password" id="confirmPassword" name="confirmPassword"
-                           placeholder="${model.data.confirmPasswordPlaceholder}" required class="input input-bordered">
+                    <div class="flex flex-col gap-1.5">
+                        <label for="confirmPassword" class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${model.data.confirmPasswordLabel}</label>
+                        <input type="password" id="confirmPassword" name="confirmPassword"
+                               placeholder="${model.data.confirmPasswordPlaceholder}" required class="input input-bordered w-full">
+                    </div>
 
-                    <button type="submit">${model.data.submitLabel}</button>
+                    <button type="submit" class="btn btn-primary w-full mt-2">${model.data.submitLabel}</button>
                 </form>
-            </div>
-            <div id="auth-result-slot" class="mt-4"></div>
-        </section>
+                <div id="auth-result-slot" class="mt-4"></div>
+            </section>
+        </div>
     </div>
 `)
