+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/SettingsPage.kte
@@ -3,10 +3,9 @@
 @param model: Page<SettingsPage>
 
 @template.io.github.rygel.outerstellar.platform.web.LayoutRouter(shell = model.shell, content = @`
-    <div style="max-width: 700px; margin: 2rem auto;">
-        <h1 style="margin-bottom: 1.5rem;">${model.data.title}</h1>
+    <div class="max-w-3xl mx-auto">
+        <h1 class="text-3xl font-extrabold mb-6">${model.data.title}</h1>
 
-        <%-- Tab bar --%>
         <nav class="tabs tabs-bordered mb-6">
             @for(tab in model.data.tabs)
                 <a href="${tab.url}"
@@ -19,8 +18,7 @@
             @endfor
         </nav>
 
-        <%-- Tab content panel --%>
-        <section class="card bg-base-200 border border-base-300" id="settings-content">
+        <section class="card bg-base-200 border border-base-300 p-6" id="settings-content">
             @if(model.data.activeTab == "profile")
                 ${model.data.profileDescription}
             @elseif(model.data.activeTab == "password")
