+++ platform-web/src/main/jte/io/github/rygel/outerstellar/platform/web/components/PageHeader.kte
@@ -7,7 +7,7 @@
 <div class="flex justify-between items-start gap-4 mb-4 p-5 bg-base-200 border border-base-300 rounded-lg">
     <div class="flex flex-col gap-1">
         @if(breadcrumb.isNotEmpty())
-            <p class="topbar-label">${breadcrumb}</p>
+            <p class="text-[0.7rem] uppercase tracking-widest font-extrabold opacity-40">${breadcrumb}</p>
         @endif
         <h1>${title}</h1>
         @if(description.isNotEmpty())
